+++ CHANGELOG.md
@@ -14,9 +14,11 @@
 - Added toast component in actions @nzambello
 - Added translations to italian @nzambello
 - Add table tile @robgietema
+- Add image sidebar @sneridagh @gomez
 - Add delete file button to file Widget @jackahl
 - Add link redirect view @robgietema
 - Add proper unique id to all fields, based on the fieldset and field name @sneridagh
+- Add QueryString widget @robgietema @sneridagh
 
 ### Changes